requirements.txt
flask
flask-cors
sentence-transformers
numpy
Pillow
requests 